tests: Extract fixture helpers from utils
